Add console parameters for easier configuration and document their usage
While the scripts are logicaly slit it is definetly not intuetive to understand
when or how to use them.
To allow for dynamic usage and to make it easier for the users all component
based scripts are given per-component command-line flags.
The make-sdcard script is given the needed ability to generate both pysical sdcard
as well images.
All of these behavious are documented in the README.md

Signed-off-by: Mimoja <mimoja@meticuloushome.com>