feat: add meticulous-brightness service to root filesystem
This commit updates the make-rootfs.sh script to install the meticulous-brightness.service
and create the appropriate symbolic link in the systemd directory. Additionally, it
includes the meticulous-brightness.service file to be tracked in the repository.

Chose to implement the default brightness logic within the service because it was simpler
and more efficient for me. Setting up the DTS was complicated, and incorporating this
functionality into the service allowed for a quicker implementation.