fix(media_handler): usb mass storage handle for recovery update is working
The purpose of media_handler.sh now is just to start the usb-rauc-install.service
and provision it the name of the device that triggered the udev rule, with this
the udev rule executable does not have to mount and look for files before, as it is
now done by the service itself. If there is a recovery update going on, the script
will not try to start the service, so it will just "reject" any new usb drive

This script now is in charge of looking for the rauc bundle and deciding to start a
recovery update or just notify the backend of the new media. It can look for the bundle
in any partition of the plugged in device. If by any reason, there is bundles in more
than one partition, it will install the first bundle found -from the partition with the
lowest number-

Now it catches the name of the whole device instead of only the partitions, in fact now it
ignores the partitions, so it is only triggered once for every device despite the its number
of partitions