meticulous-brightness.service: Don't hardcode the max brightness
The usage of a max brightness of 4096 requires us to always have these
brightness levels available now and in the future.
As we will be migrating to different kernels with different backlight
implementations in the future we shouldn't hardcode these values any
longer.
NOTE: This should also be implemented in the backend where the
brightness for idle is also hardcoded

Signed-off-by: Mimoja <mimoja@mimoja.de>